+++ docs/source/changelog.rst
@@ -3464,7 +3464,12 @@ Current C++/SQLite client, Python/SQLAlchemy server
 - **New task:** :ref:`CPFT Research Preferences  <cpft_research_preferences>` (Database revision 0067).
 - **New task:** :ref:`CPFT Post-Covid Clinic Medical Questionnaire  <cpft_covid_medical>` (Database revision 0068).
 
+
 .. _changelog_v2_4_10:
+
 
 **Client and server v2.4.10, IN PROGRESS**
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Any user with the right privilege (not just the group administrator) can
+  add/edit/delete and send emails to patients created on the server. The menu option **Manage scheduled tasks for patients** is now **Manage patients and their tasks**. (Database revision 0069)